# docs/openapi/pbb-hotline-beta.yaml
openapi: 3.1.0
info:
  title: PBB Hotline Beta API
  version: 0.1.0-draft
  summary: Phase 1 API draft for PBB Hotline Beta aligned to the current inventory.
servers:
  - url: https://hotline.pbb.ph
...
paths:
...
  /api/caller/incidents/{incident}:
    get:
      tags: [Caller]
      summary: Get one caller-visible incident view
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Caller incident payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
...
  /api/operator/incidents:
    get:
      tags: [Operator]
      summary: Get operator incident list feed
      security:
        - sessionAuth: []
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
          description: Comma-separated incident statuses, such as `active,deferred`.
      responses:
        '200':
          description: Operator incident list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperatorIncidentListResponse'
  /api/operator/incidents/{incident}:
    get:
      tags: [Operator]
      summary: Get full operator workbench payload
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Incident workbench payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '403':
          description: Operator may no longer open this incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident no longer available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/incidents/{incident}/messages:
    get:
      tags: [Messaging]
      summary: Get incident chat history
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Incident messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
    post:
      tags: [Messaging]
      summary: Create incident message during live call
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMessageRequest'
      responses:
        '200':
          description: Message created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '409':
          description: Messaging currently blocked by live-call rules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: laravel_session
  parameters:
    IncidentId:
      in: path
      name: incident
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
        code: { type: string, nullable: true }
        details:
          oneOf:
            - type: object
            - type: array
              items: {}
            - type: string
          nullable: true
      required: [message]
...
    CallerSummary:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        avatar: { type: string, nullable: true }
        mobile: { type: string }
      required: [id, name, mobile]
    OperatorSummary:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        level: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
      required: [id, name]
...
    CallParticipant:
      type: object
      properties:
        id: { type: integer }
        call_session_id: { type: integer }
        user_id: { type: integer }
        participant_role:
          type: string
          enum: [caller, operator]
        joined_at: { type: string, format: date-time }
        left_at: { type: string, format: date-time, nullable: true }
      required: [id, call_session_id, user_id, participant_role, joined_at]
    CallSession:
      type: object
      properties:
        id: { type: integer }
        incident_id: { type: integer }
        caller_id: { type: integer }
        status:
          type: string
          enum: [calling, in_progress, ended]
        outcome:
          type: string
          enum: [answered, timed_out, declined_by_operator, cancelled_by_caller, ended_by_operator, ended_by_caller]
        started_at: { type: string, format: date-time }
        answered_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        participants:
          type: array
          items:
            $ref: '#/components/schemas/CallParticipant'
      required: [id, incident_id, caller_id, status, outcome, started_at, created_at, updated_at]
    Transfer:
      type: object
      properties:
        id: { type: integer }
        incident_id: { type: integer }
        from_operator_id: { type: integer }
        to_operator_id: { type: integer }
        reason: { type: string }
        status: { type: string }
        requested_at: { type: string, format: date-time }
        accepted_at: { type: string, format: date-time, nullable: true }
        rejected_at: { type: string, format: date-time, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
      required: [id, incident_id, from_operator_id, to_operator_id, reason, status, requested_at]
    MessageAttachment:
      type: object
      properties:
        id: { type: integer }
        message_id: { type: integer }
        type: { type: string }
        mime_type: { type: string }
        original_filename: { type: string }
        stored_path: { type: string }
        file_size: { type: integer }
        thumbnail_path: { type: string, nullable: true }
        uploaded_by: { type: integer }
        created_at: { type: string, format: date-time }
      required: [id, message_id, type, mime_type, original_filename, stored_path, file_size, uploaded_by, created_at]
    Message:
      type: object
      properties:
        id: { type: integer }
        incident_id: { type: integer }
        sender_id: { type: integer }
        sender_role:
          type: string
          enum: [caller, operator, system]
        sender_name: { type: string }
        sender_avatar: { type: string, nullable: true }
        body: { type: string }
        type:
          type: string
          enum: [message, system]
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/MessageAttachment'
        created_at: { type: string, format: date-time }
      required: [id, incident_id, sender_id, sender_role, sender_name, body, type, attachments, created_at]
    MediaRecord:
      type: object
      properties:
        id: { type: integer }
        incident_id: { type: integer }
        call_session_id: { type: integer }
        type:
          type: string
          enum: [audio_peer, caller_video]
        peer_user_id: { type: integer, nullable: true }
        peer_role: { type: string, nullable: true }
        peer_label: { type: string, nullable: true }
        path: { type: string }
        duration_seconds: { type: integer, nullable: true }
        metadata:
          type: object
          additionalProperties: true
        created_at: { type: string, format: date-time }
        available_at: { type: string, format: date-time, nullable: true }
      required: [id, incident_id, call_session_id, type, path, created_at]
    IncidentResourceNeeded:
      type: object
      properties:
        resource_type_id: { type: integer }
        resource_name: { type: string }
        quantity_required: { type: integer }
        notes: { type: string, nullable: true }
      required: [resource_type_id, resource_name, quantity_required]
    IncidentTypeDetail:
      type: object
      properties:
        id: { type: integer }
        incident_id: { type: integer }
        incident_type_id: { type: integer }
        field_id: { type: integer }
        field_label: { type: string }
        field_key: { type: string }
        field_value: {}
        input_type: { type: string }
        options:
          type: array
          items: {}
        unit: { type: string, nullable: true }
        placeholder: { type: string, nullable: true }
        is_required: { type: boolean }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, incident_id, incident_type_id, field_id, field_label, field_key, field_value, input_type, is_required, sort_order, created_at, updated_at]
    TeamResourceInventory:
      type: object
      properties:
        id: { type: integer }
        team_id: { type: integer }
        resource_type_id: { type: integer }
        resource_name: { type: string }
        quantity_available: { type: integer }
        created_at: { type: string, format: date-time }
      required: [id, team_id, resource_type_id, resource_name, quantity_available, created_at]
    Team:
      type: object
      properties:
        id: { type: integer }
        team_category_id: { type: integer, nullable: true }
        category: { type: string, nullable: true }
        name: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        resources:
          type: array
          items:
            $ref: '#/components/schemas/TeamResourceInventory'
      required: [id, name, status, created_at, resources]
    TeamAssignment:
      type: object
      properties:
        id: { type: integer }
        incident_id: { type: integer }
        team_id: { type: integer }
        team: { $ref: '#/components/schemas/Team' }
        assigned_by_operator_id: { type: integer }
        status:
          type: string
          enum: [Assigned, Requested, Accepted, En-route, On-Scene, Completed, Cancelled]
        contact_person: { type: string, nullable: true }
        cancelled_from_status: { type: string, nullable: true }
        cancel_reason_code: { type: string, nullable: true }
        cancelled_by_operator_id: { type: integer, nullable: true }
        allocated_resources:
          type: array
          items:
            type: object
            properties:
              resource_type_id: { type: integer }
              quantity_allocated: { type: integer }
            required: [resource_type_id, quantity_allocated]
        assigned_at: { type: string, format: date-time, nullable: true }
        accepted_at: { type: string, format: date-time, nullable: true }
        enroute_at: { type: string, format: date-time, nullable: true }
        arrived_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        cancelled_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, incident_id, team_id, team, assigned_by_operator_id, status, allocated_resources, created_at, updated_at]
    Incident:
      type: object
      properties:
        id: { type: integer }
        display_id: { type: string }
        caller_id: { type: integer }
        caller: { $ref: '#/components/schemas/CallerSummary' }
        actual_caller_name: { type: string }
        actual_caller_relationship: { type: string }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        location: { type: string, nullable: true }
        location_road: { type: string, nullable: true }
        location_suburb: { type: string, nullable: true }
        location_barangay: { type: string, nullable: true }
        location_citymunicipality: { type: string, nullable: true }
        location_country: { type: string, nullable: true }
        operator_id: { type: integer }
        operator: { $ref: '#/components/schemas/OperatorSummary' }
        status:
          type: string
          enum: [Active, Deferred, Discarded, Resolved]
        alert_level:
          type: string
          enum: [Normal, Elevated, Critical]
        called_at: { type: string, format: date-time, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        other_details: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        call_history:
          type: array
          items:
            $ref: '#/components/schemas/CallSession'
        transfer_history:
          type: array
          items:
            $ref: '#/components/schemas/Transfer'
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        media:
          type: array
          items:
            $ref: '#/components/schemas/MediaRecord'
        incident_type_details:
          type: array
          items:
            $ref: '#/components/schemas/IncidentTypeDetail'
        incident_resources_needed:
          type: array
          items:
            $ref: '#/components/schemas/IncidentResourceNeeded'
        team_assignments:
          type: array
          items:
            $ref: '#/components/schemas/TeamAssignment'
      required: [id, display_id, caller_id, caller, actual_caller_name, actual_caller_relationship, operator_id, operator, status, alert_level, created_at, updated_at, call_history, transfer_history, messages, media, incident_type_details, incident_resources_needed, team_assignments]
    IncidentListItem:
      type: object
      properties:
        id: { type: integer }
        display_id: { type: string }
        caller_avatar: { type: string, nullable: true }
        actual_caller_name: { type: string }
        status:
          type: string
          enum: [Active, Deferred, Discarded, Resolved]
        created_at: { type: string, format: date-time }
      required: [id, display_id, actual_caller_name, status, created_at]
...
    OperatorIncidentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IncidentListItem'
      required: [items]
...
    CreateMessageRequest:
      type: object
      properties:
        body: { type: string }
      required: [body]
...
    MessageListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
      required: [items]
...